Refactored spec helpers and move specs to correct spec class
