+++ lib/streamy/helpers/message_parser.rb
@@ -3,20 +3,18 @@ require "avro_turf/messaging"
 module Streamy
   module Helpers
     module MessageParser
-      def hashify_messages(message_bus_deliveries, encoding_type)
+      def hashify_messages(message_bus_deliveries)
         message_bus_deliveries.map do |message|
           message_hash = message.dup
-          message_hash[:payload] = parse_message(message_hash[:payload], encoding_type)
+          message_hash[:payload] = parse_message(message_hash[:payload])
           message_hash
         end
       end
 
-      def parse_message(message_payload, encoding_type)
-        if encoding_type == :avro
-          avro.decode(message_payload).deep_symbolize_keys
-        else
-          JSON.parse(message_payload).deep_symbolize_keys
-        end
+      def parse_message(message_payload)
+        JSON.parse(message_payload).deep_symbolize_keys
+      rescue JSON::ParserError
+        avro.decode(message_payload).deep_symbolize_keys
       end
 
       def avro
