+++ test/event_test.rb
@@ -2,36 +2,6 @@ require "test_helper"
 
 module Streamy
   class EventTest < Minitest::Test
-    class ValidEvent < Event
-      def topic; end
-      def body; end
-      def event_time; end
-    end
-
-    class ValidEventWithParams < Event
-      def initialize(i)
-        @i = i
-      end
-
-      def body
-        {
-          i: @i
-        }
-      end
-
-      def event_time
-        -"now"
-      end
-
-      def topic
-        -"valid_event_with_params_topic"
-      end
-    end
-
-    class OveriddenPriority < ValidEvent
-      priority :low
-    end
-
     class EventWithoutTopic < Event
       def event_time; end
       def body; end
@@ -63,43 +33,6 @@ module Streamy
       assert_runtime_error "body must be implemented on Streamy::EventTest::EventWithoutBody" do
         EventWithoutBody.publish
       end
-    end
-
-    def test_default_priority
-      ValidEvent.publish
-
-      assert_published_event(priority: :standard)
-    end
-
-    def test_overidden_priority
-      OveriddenPriority.publish
-
-      assert_published_event(priority: :low)
-    end
-
-    def test_deliver_batched_events_in_block
-      ValidEventWithParams.deliver do |event|
-        2.times do |i|
-          event.publish(i)
-        end
-      end
-
-      assert_published_event(
-        topic: "valid_event_with_params_topic",
-        payload: {
-            type: "valid_event_with_params", 
-            body: { i: 0 }, 
-            event_time: "now"
-          }
-      )
-      assert_published_event(
-        topic: "valid_event_with_params_topic",
-        payload: {
-            type: "valid_event_with_params", 
-            body: { i: 1 }, 
-            event_time: "now"
-          }
-      )
     end
   end
 end
