+++ test/json_event_test.rb
@@ -19,6 +19,30 @@ module Streamy
       end
     end
 
+    class OveriddenPriority < TestEvent
+      priority :low
+    end
+
+    class ValidEventWithParams < TestEvent
+      def initialize(event_number)
+        @event_number = event_number
+      end
+
+      def topic
+        -"valid_event_with_params_topic"
+      end
+
+      def body
+        {
+          event_number: @event_number
+        }
+      end
+
+      def event_time
+        -"now"
+      end
+    end
+
     def test_publish
       SecureRandom.stubs(:uuid).returns("IAMUUID")
 
@@ -31,7 +55,44 @@ module Streamy
           type: "test_event",
           body: { smoked: "true", streaky: "false" },
           event_time: "nowish"
-        }.to_json
+        }
+      )
+    end
+
+    def test_default_priority
+      TestEvent.publish
+
+      assert_published_event(priority: :standard)
+    end
+
+    def test_overidden_priority
+      OveriddenPriority.publish
+
+      assert_published_event(priority: :low)
+    end
+
+    def test_deliver_batched_events_in_block
+      ValidEventWithParams.deliver do |event|
+        2.times do |i|
+          event.publish(i)
+        end
+      end
+
+      assert_published_event(
+        topic: "valid_event_with_params_topic",
+        payload: {
+          type: "valid_event_with_params",
+          body: { event_number: 0 },
+          event_time: "now"
+        }
+      )
+      assert_published_event(
+        topic: "valid_event_with_params_topic",
+        payload: {
+          type: "valid_event_with_params",
+          body: { event_number: 1 },
+          event_time: "now"
+        }
       )
     end
   end
